Sine and cosine trigonometric functions program
This program obtains the trigonometric functions of sine and cosine through the Taylor power series and the tangent function by dividing the result of sine and cosine. Various aspects have been modified such as: the documentation contained in comments; the tangent function was added; the calculation methods for sine and cosine were explained; added an enumeration to access the general sine and cosine function, among many other tweaks.