Program that generates Pascal's Triangle
A C++ Program is added that generates two Pascal's Triangles using some of the key concepts of Object Oriented Programming in that language. This program shows each row of the triangle in question in such a way that the row, the column and the generated value of the calculation are reported.