Program that generates Pascal's Triangle
This program generates a Pascal's Triangle and uses different Object Oriented Programming methodologies. Methods have been added to get and set the value of the number of rows of the Pascal's Triangle to be instantiated. A new method was added to clear the attributes, both of the number of lines and the pointer of pointers that hosts the Pascal's Triangle. Some new information messages have been added and this is considered a more complete and stabilized version.