Dynamic Multiplication Tables Program
This simple program is in charge of creating objects of the 'My List' class dynamically and, in turn, saving them in C++ dynamic lists, in such a way that each object of the 'My List' type is created, values are assigned to its internal attribute of type list, and is often stored in another list in order to destroy the current instance. In the end, the list is left with objects of type 'My List' and that can be executed and accessed by means of iterators.