Dynamic Multiplication Tables Program
This program uses two types of collectors: the list and the vector; Its purpose is to store in these collectors an object of type 'My List' that also has its own collector of type pointer to list, with the aim of learning to use dynamic structures from the C++ STL library. The 'size_t' data type is also used.