Dynamic Multiplication Tables Program
In this program for generating dynamic multiplication tables, constant references were added to certain methods so that their formal parameter cannot be modified or altered within the same method that reads it, as well as the same type of references but to the methods that read or access the attributes of the 'My List' class to ensure that they will not be changed in the future. This same constant reference is used in the function that only prints the values contained in the 'MyList' class as far as its formal parameter is concerned.