Pascal's Triangle Program Generator
This program has included some restrictions both in its methods so that they are constants and prevent them from modifying the attributes of the instantiated class, and some of the formal parameters they receive are constants. It is recommended to evaluate in the future if the method that sets the number of lines that each instance of a triangle should have should be protected, because if the report methods are called, an offset could occur in the reading of the pointer of pointers that it stores the 'Pascal's Triangle'.