Fibonacci Number Generator Program
This program stores each record generated from the series of Fibonacci numbers in a structure vector, informing the index number, the first and second referenced values and their sum, dragging, of course, that the first value takes the value of the second. and the second the value of the sum of the previous ones.