Tangent trigonometric function program
The purpose of this program, although with deficiencies in the calculation, is to obtain the tangent of any number expressed in degrees and through the series of numbers up/down and the Bernoulli numbers to obtain the trend of the same function, although with a divergent value. This function is limited by the number of maximum iterations because C language variables are overflowed by huge products obtained from extreme powers and factorials. It is a function that requires adjustments, improvements and it is not convenient to take it as a general function for the tangent of any number.