Pascal's Triangle Program Generator
This program protected the method that sets the number of lines of the Pascal Triangle pointer pointer, and the option to set it with public visibility was eliminated, with the aim of improving security in dynamic memory management. It should be clarified that if the entered value is less than the previous number of rows, memory segment violations may be caused.