Pascal's Triangle Program Generator
In this program we returned to the function that sets the number of rows of Pascal's Triangle to public visibility. Added tracking messages for each dynamically created instance. The menu was modified so that only when creating a new Pascal's Triangle within the current instance is it requested to capture the number of rows. Note that requesting a different number of rows than was used to create the Triangle object's pointer pointer can cause memory segfaults in other called functions.