Pascal's Triangle Menu Generator Program
A friendly function was added to this program to access the private and protected elements of the instantiated class 'Pascal's Triangle'. The options that can be executed on said instantiated class were improved, and the display of the action menu was improved, as well as new class manipulation options were inserted. Added pass-by-reference of the instantiated Pascal's Triangle to the unit testing function and changed some user information messages.