Pascal's Triangle Menu Generator Program
Functions were added to this program both for the main call of the unit tests by the user, as well as a function to return a vector of vectors with the coefficients calculated and stored in the pointer of pointers of the instantiated class of Pascal's Triangle. There are still issues regarding memory segment violations due to read exceedances of the aforementioned double pointer. New options were also added to the interactive menu and some user warning and caution messages were changed.