Pascal's Triangle Menu Generator Program
The constant return types that some of its functions had were removed from this program, and the user interaction menu was improved. The number of instances that the user can create is restricted to avoid the tedium of exiting the treatment of each and, last but not least, memory segment violations continue to occur in certain scenarios when the class attribute of the number of rows, since it is common to all its methods of the 'Pascal's Triangle' class.