Pascal's Triangle Menu Generator Program
This program, which generates instances at the user's request of a Pascal's Triangle, has been stabilized in such a way that when a new value is requested to be entered for the number of rows of the mentioned Triangle, the previous memory regions are first cleaned and set the new value to the corresponding attribute of the class. So far it has been tested, and there have been no memory segment violations. It remains to validate the data entry when the number of lines is requested, as it presents some errors.