Prime number generator program
The purpose of this program is to generate a sequence of prime numbers from one to the value 'n'. It is understood that a prime number is one that only admits itself and unity as unique divisors. Consequently, if a number has other divisors apart from those already mentioned, it is considered a composite number.