Pascal's Triangle generator program
This program was only refactored with the condition that the pointer of pointers that stores Pascal's Triangle, instead of not being equal to null, is simply checked intrinsically by the compiler if it is true or not. When true it has a direction and when false it has the value of absolute zero.