C++ TLS Collection Generator Program
Some corrections were made to this program such as modification of orientation comments, the inclusion of vectors that, apart from the lists, store 'My List' type objects and that have a separation on the screen for each dump of values. It must be understood that lists and vectors contain objects of type 'My List' which in turn in their own attributes contain lists and vectors with unsigned integer numerical values.