Hanoi Towers Algorithm Solving Program
This program solves the Towers of Hanoi algorithm through recursive cycles that solve the enigma of moving a certain number of saucers from one container to another, passing through two other containers, in such a way that the discs are stacked from the smallest to the largest starting through the upper part and that each transfer it is checked that a larger disc can never be on a smaller disc. The algorithm requires (2^n) iterations to solve.