Random number timer program
To this program, some more messages were added to the screen about latency and PC clock pulses, as well as reformatting the output to make it more understandable to the user. Likewise, the code was refactored and the use of new data types was expanded for use in this program.