Positive and negative powers generator program
The purpose of this program is to raise a base number to a certain power, whether positive or negative. In the first case, successive multiplications are made on the same basis; In the second case, successive divisions are made on the same basis. There are two implementations to enhance a number in this program, the first is eminently self-recursive, and the second uses conditional loops.