Mathematical Functions General Project
Some modifications have been made to the program that generates Pascal's Triangles, and refactorings and re-optimizations of the source code were made. The issue to discuss is why once a dynamic object is created, stored in a vector, and then destroyed to create a new one, the destructors are called multiply for each call.