Mathematical Functions General Project
I made these programs to practice the simplest dynamic data structures that are stacks and queues, using pointers, in order to understand their operation and use in the future. These programs lack comments and the nomenclature of their variables is quite brief. I tried to be as optimized as possible in its algorithms.