Mathematical Functions General Project
More output messages were incorporated into this program, both from different ways of invoking the memory addresses occupied by each element of Pascal's Triangle and from the different ways of referring to the value stored in the pointer of pointers, that is, they were varied. the notations of both arrays and pointers in Pascal's Triangle.