Mathematical Functions General Project
This brief example in C++ is responsible for exemplifying the use of arguments by value and by reference in functions within namespaces, in order to learn how to use them to convert a date and time into long integer formats and break them down into their respective units. logical independents.