Mathematical Functions General Project
New features were added to this program to request the name of a file to save the reports or use a default name, as well as modify a method that obtains a response from the user by customizing its input message. So far, it's a pretty functional program that generates Pascal's Triangles.