Mathematical Functions General Project
Small modifications were added to this program for certain functions that previously did not return any value and now return at least one integer value of the elements processed in each intervened function. Also made a change so that the default file name is changed or left depending on the answer given to a regeneration question.