Mathematical Functions General Project
This is a program that simulates the typical game of Cat, 'Tic Tac Toe' or Tic Tac Toe, and its purpose is for three squares in a line to be occupied by the same symbol. It is a program that requires a lot of optimization because I based it on a C++ video tutorial to develop it.