Mathematical Functions General Project
A new display value was added to this Pascal's Triangle program, which consists of multiple expressions that mutually share arrays and pointers in C and C++, with the purpose of exemplifying access to both memory regions and data in multiple ways. the values of the pointer of pointers that each dynamically generated Pascal's Triangle stores.