Mathematical Functions General Project
The purpose of this program is to perform the typical game of rock, paper, scissors in C++ language that I extracted from a certain video on YouTube to inspire me. It must be clarified that this program needs to be debugged because it has a lot of repeated code and must be optimized.