Mathematical Functions General Project
This program validates both credit and debit card numbers through the Luhn Algorithm, which is responsible for adding the digits in odd positions as they appear, and adding the odd digits by doubling the value obtained and separating the new digits. The sum of even and odd digits when divided by ten must have a remainder of zero to be considered a valid card number.