Mathematical Functions General Project
This program had its code refactored to add the digits in even and odd positions of the card number using a new index and checking if the position is even or odd. In this way, more results are returned such as the analyzed digit number, its respective value, the length of the number of digits on the card and the total number of digits added in even and odd positions of the analyzed card.