Mathematical Functions General Project
Only some adjustments were made to this program in the comments to better explain the calculations to obtain the trigonometric functions of the sine and cosine through the Taylor series, and to change some formal parameters of the functions that are responsible for their respective calculations. mathematicians. Also added some new output messages.