Mathematical Functions General Project
This program validates a given date that was entered by year, month and day. It verifies that it is within the logical limits for each of these variables, and if it is correct, a date formatted with the name of the day and month corresponding to said date is returned.