Mathematical Functions General Project
2025/03/14 - 21:19:24.
This program validates a given date and time, determining respectively its year, month and day, and if correct, it calculates Easter Sunday of that year, the current Julian year, the name of the day of the week that corresponds to it as well as the total number of days elapsed since January 1, 1582. In the case of time, its hour, minute and second are requested respectively, and if correct, it calculates the total number of seconds that correspond to that given time.