Mathematical Functions General Project
03/14/2025 - 9:19:24 PM
This program validates a given date and time, determining their year, month, and day, respectively. If correct, it calculates Easter Sunday of that year, the current Julian year, the name of the corresponding day of the week, the total number of days elapsed since January 1, 1582, and its inverse, which is to convert those accumulated days into their respective date (day, month, and year). For the time, the hour, minute, and second are requested, respectively. If correct, the total number of seconds corresponding to that given time is calculated.