Mathematical Functions General Project
This program validates a given date and time, determining their year, month, and day, respectively. If correct, it calculates Easter Sunday of that year, the current Julian year, the remaining days of the year starting from the detected Julian year, the name of the corresponding day of the week, the total number of days elapsed since January 1, 1582, and its inverse function, which converts those accumulated days into their respective date (day, month, and year). For the time, the hour, minute, and second are requested, respectively. If correct, the total number of seconds corresponding to that given time is calculated.