Mathematical Functions General Project
The purpose of this program is to implement a program that creates a sequential formatted text file and retrieves and copies it to another file of the same type, even setting limits on the number of records to be copied. This program is ideal if you need to copy a certain number of records from one file to another, or simply create a blank one.