Mathematical Functions General Project
The purpose of this program is to determine the natural logarithm to base 'e' of a given base value through Taylor series calculation cycles, thereby obtaining an approximation of that value. This program takes a while to calculate due to the large number of powers and factorials it must perform. The program is computationally intensive; it takes a few minutes to obtain the result, even though it has been simplified.