Mathematical Functions General Project
The purpose of this program is to search for a given value as a key within an array of integers using traditional binary search, recursive binary search, binary search with a 'delta' adjustment factor, approximate interpolation locating, sequential search with starting position and scanning search with start position and address values of each value looked, through a menu of options. If said key value is found, it returns its relative position or entry number (displacement index) within said numeric array.