Mathematical Functions General Project
This program exemplifies different techniques of Object Oriented Programming in C++ language given a class of type Point2D that has 2 attributes 'x' and 'y' and its setter and getter methods, which inherits another class called Point3D with its attribute 'z' and its own setter and getter methods. Different classes of the 'std' namespace are overloaded and templates and friend classes are used, as well as everything is contained in a given namespace.