Mathematical Functions General Project
The primary purpose of this program is to generate several three-dimensional and two-dimensional matrices, and a one-dimensional array. The first is to store a series of matrices, each with a variable number of rows, and each with a variable number of columns or cells. The two-dimensional matrix stores the number of rows that each three-dimensional matrix number has individually per entry. Finally, the one-dimensional array stores the number of columns that each row has, individually determined from each of the matrices determined by the user. This is a great example of highly variable dynamic memory allocation for each content captured by the user.