Mathematical Functions General Project
The purpose of this program is to illustrate the use of function pointers with a variable number of arguments, such that said function pointer can be called with different lists of formal parameters and return the same return type. In this way, a function is also called that receives as arguments all the formal parameters of the functions that can be called by pointer, the function pointer itself, and an enumeration of the possible calling options.