Mathematical Functions General Project
The purpose of this program is to create an array of variables of type 'void *' that store any type of value. These variables are sent to a function that accepts variable arguments, and then sum them up depending on the type of value in question. It is being tested because the values ​​are first stored in an array, and each value is deconverted according to its type. In reality, since this is a function that accepts variable arguments, only one test argument is sent in the syntax of its call.