Mathematical Functions General Project
The purpose of this program is to create a chessboard on which up to eight non-attacking queen pieces are randomly placed on the board, given a random seed for piece placement. Typically, the number of queens is between 5 and 7, placed regularly. Cases where eight queen pieces are placed are rare.