Mathematical Functions General Project
This program uses a base class called 'MyPoint' that inherits from a derived class called 'MyPoint3D' and uses polymorphism, dynamic memory allocation, various calling nomenclatures for each of the methods and operators of both classes, as well as instantiating them with the desired data type since they are template-type classes, and even performing dynamic exchanges to access virtual and overloaded synonymous methods that are shared by both classes. It is a good example of the use of various Object-Oriented Programming paradigms with the C++ language.