Mathematical Functions General Project
This program aims to efficiently demonstrate the popular Rock-Paper-Scissors game, commonly played in English-speaking countries.  It uses various data structures and functions in C++ to determine the possible outcomes for both the user and the computer. The computer player requires a random seed number to initialize its random number generator, which is custom-programmed to ensure a certain level of unpredictability in its moves.

Rename RockPapS.cpp to RockPaperScissors.cpp.