Mathematical Functions General Project
The primary purpose of this program is to demonstrate the use of file pointers in ANSI C. It shows how to create a file with formatted records, retrieve and display its contents, and copy those records to another file.  Finally, it displays the contents of the new file and compares the records in both files to verify that they are identical, except for the encoding format.