Mathematical Functions General Project
The primary purpose of this program is to use some of the most important functions of the C++ STL library for my personal practice. This program creates three doubly linked lists, generating the first with random numbers from one to ten. These lists are then placed in two lists: one with numbers from one to five, and the second with numbers from six to ten. This example is just a glimpse into my advanced C++ learning process.