Mathematical Functions General Project
A special function was added to this program to search for a certain value in the binary search tree and some informative messages were added that in a binary search tree there should not be duplicate or identical values.
The primary purpose of this program is to generate a binary search tree and perform operations such as adding nodes, searching for a specific node, deleting the entire tree, and displaying all nodes in the tree, depending on whether they are in Preorder, Inorder, Postorder, or their inverses. The deletion of a specific node from the binary search tree is still missing.