Add files via upload
This program aims to calculate the number of bags containing a specific amount of candy for a chimpanzee and reduce the number of bags to a specific quantity by removing the previous one and, using a delta factor, accumulating the remaining bags until a total amount of candy consumed by all the bags is reached. It should be noted that this is a renewed version for personal learning purposes.